Add --environment-config param
Basically, I wanted to set lstrip_blocks and trim_blocks to true when using
j2cli. I didn't see a way to do that, so this is a generic way to achieve it.
You can now pass a config file which use kwargs-expanded and passed to the
jinja2 Environment. This allows the user to break things confusingly, but if
they are using this feature, presumably they already understand jinja2 well
enough to debug the problem for themself.