Set a maximum version on jinja
At the moment, the latest version will always be fetched. However, as of version 3.0.0, python 2.7 support
was dropped, meaning there is a BC issue since pipelines still using python 2.7 are breaking.

Unless this is an deliberate decision, a max version should be set on the jinja requirement and future changelogs should
state that python 2.7/3.5 support is dropped if it were to be removed again